Rename ping subpackage to status
