+++ datahub/status/urls.py
@@ -1,6 +1,6 @@
 from django.conf.urls import url
 
-from datahub.ping.views import version
+from datahub.status.views import version
 
 urlpatterns = [
     url(r'^version', version, name='version'),
